+++ packages/altive_lints_plugin/test/src/utils/files_utils_test.dart
@@ -36,17 +36,14 @@ void main() {
       expect(isTestFile(file), isFalse);
     });
 
-    test(
-      'False if the path or directory contains the string "test" '
-      'but it is not an exact match',
-      () {
-        final resourceProvider = PhysicalResourceProvider.INSTANCE;
+    test('False if the path or directory contains the string "test" '
+        'but it is not an exact match', () {
+      final resourceProvider = PhysicalResourceProvider.INSTANCE;
 
-        const filePath = '/project/testable/testable_file.dart';
-        final file = resourceProvider.getFile(filePath);
-        expect(isTestFile(file), isFalse);
-      },
-    );
+      const filePath = '/project/testable/testable_file.dart';
+      final file = resourceProvider.getFile(filePath);
+      expect(isTestFile(file), isFalse);
+    });
 
     test('False if it is in lib directory and file name is test.dart', () {
       final resourceProvider = PhysicalResourceProvider.INSTANCE;